teste loopback
testando enviar informaçoes e ver se recebe certo

+++ PAINEL_CODIGO/TouchGFX/gui/src/model/Model.cpp
@@ -6,6 +6,7 @@
 #include <stdint.h>
 #include "cmsis_os2.h"
 #include "can_types.h"
+#include <cstring>
 
 extern "C" {
     #include "main.h" // Para garantir que can_msg_t seja reconhecida
@@ -95,25 +96,12 @@ void Model::tick()
         {
             model_recebeu_fila++; // Se este número subir, o Model está lendo a fila!
 
-            // vou fazer de maneira poluida primeiro pq é mais provavel de funcionar, depois
-            // de garantir que funciona ajusto pra deixar bonito
-//            uint16_t rpm = (msg_recebida.data[0] << 8 | msg_recebida.data[1]);
-//            uint16_t TempMotor = (msg_recebida.data[2] << 8 | msg_recebida.data[3]);
-//            uint16_t TempInversor = (msg_recebida.data[6] << 8 | msg_recebida.data[7]);
-//            uint16_t freio = msg_recebida.data[5];
-//            uint16_t acelerador = msg_recebida.data[4];
-//            uint16_t TempAcumulador = msg_recebida.data[7];
-//            uint16_t TensaoHV = memcpy(&AcumCurrent, &RxData[4], sizeof(float));
-//            uint16_t TensaoInversor = memcpy(&AcumCurrent, &RxData[0], sizeof(float));
-//            uint16_t valor = msg_recebida.data[0];
-
-
             switch (msg_recebida.id)
             {
             case 0x121: {
-                            uint16_t freio = msg_recebida.data[5];
-                            uint16_t acelerador = msg_recebida.data[4];
-                            uint16_t TempAcumulador = msg_recebida.data[7];
+                            uint8_t freio = msg_recebida.data[5];
+                            uint8_t acelerador = msg_recebida.data[4];
+                            uint8_t TempAcumulador = msg_recebida.data[7];
 
                             modelListener->updateFreioValue(freio);
                             modelListener->updateAceleradorValue(acelerador);
@@ -122,9 +110,9 @@ void Model::tick()
                         }
 
             case 0x420: {
-                            uint16_t rpm = (msg_recebida.data[0] << 8) | msg_recebida.data[1];
-                            uint16_t TempMotor = (msg_recebida.data[2] << 8) | msg_recebida.data[3];
-                            uint16_t TempInversor = (msg_recebida.data[6] << 8) | msg_recebida.data[7];
+                            uint16_t rpm = ((uint16_t)msg_recebida.data[0] << 8) | msg_recebida.data[1];
+                            uint16_t TempMotor = ((uint16_t)msg_recebida.data[2] << 8) | msg_recebida.data[3];
+                            uint16_t TempInversor = ((uint16_t)msg_recebida.data[6] << 8) | msg_recebida.data[7];
 
                             modelListener->updateRPMValue(rpm);
                             modelListener->updateTempMotorValue(TempMotor);
